build(Gradle): Migrate {all,sub}projects to convention plugins
Migrate logic from `allprojects` and `subprojects` to the respective
conventions plugins. This avoids the need for distinguishing non-Java
projects in `subprojects`. The `gradle-model` and `gradle-plugin` projects
accompanying the `gradle-inspector` now need to explicitly become Kotlin
projects by applying the respective conventions plugin.
