+++ build.gradle.kts
@@ -19,9 +19,6 @@
 
 import com.github.benmanes.gradle.versions.updates.DependencyUpdatesTask
 
-import io.gitlab.arturbosch.detekt.Detekt
-import io.gitlab.arturbosch.detekt.report.ReportMergeTask
-
 import org.eclipse.jgit.api.Git
 import org.eclipse.jgit.lib.Config
 import org.eclipse.jgit.lib.Constants
@@ -31,22 +28,14 @@ import org.eclipse.jgit.treewalk.TreeWalk
 import org.eclipse.jgit.util.FS
 import org.eclipse.jgit.util.SystemReader
 
-import org.gradle.api.tasks.testing.logging.TestExceptionFormat
-import org.gradle.api.tasks.testing.logging.TestLogEvent
-
 import org.jetbrains.gradle.ext.Gradle
 import org.jetbrains.gradle.ext.runConfigurations
 import org.jetbrains.gradle.ext.settings
-import org.jetbrains.kotlin.gradle.plugin.KotlinCompilation
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 plugins {
     // Apply third-party plugins.
     alias(libs.plugins.dependencyAnalysis)
-    alias(libs.plugins.detekt)
-    alias(libs.plugins.dokkatoo)
     alias(libs.plugins.ideaExt)
-    alias(libs.plugins.kotlin)
     alias(libs.plugins.versions)
 }
 
@@ -96,10 +85,6 @@ if (version == Project.DEFAULT_VERSION) {
 
 logger.lifecycle("Building ORT version $version.")
 
-// See https://kotlinlang.org/docs/compiler-reference.html#jvm-target-version.
-val javaVersion = JavaVersion.current()
-val maxKotlinJvmTarget = javaVersion.majorVersion.toInt().coerceAtMost(19)
-
 idea {
     project {
         settings {
@@ -131,292 +116,6 @@ tasks.named<DependencyUpdatesTask>("dependencyUpdates") {
 
     rejectVersionIf {
         candidate.version.matches(nonFinalQualifiersRegex)
-    }
-}
-
-val mergeDetektReports by tasks.registering(ReportMergeTask::class) {
-    output.set(rootProject.buildDir.resolve("reports/detekt/merged.sarif"))
-}
-
-allprojects {
-    repositories {
-        mavenCentral()
-    }
-
-    apply(plugin = "io.gitlab.arturbosch.detekt")
-
-    // Note: Kotlin DSL cannot directly access configurations that are created by applying a plugin in the very same
-    // project, thus put configuration names in quotes to leverage lazy lookup.
-    dependencies {
-        "detektPlugins"(project(":detekt-rules"))
-
-        "detektPlugins"("io.gitlab.arturbosch.detekt:detekt-formatting:${rootProject.libs.versions.detektPlugin.get()}")
-    }
-
-    detekt {
-        // Only configure differences to the default.
-        buildUponDefaultConfig = true
-        config.from(files("$rootDir/.detekt.yml"))
-
-        source.from(fileTree(".") { include("*.gradle.kts") }, "src/funTest/kotlin")
-
-        basePath = rootProject.projectDir.path
-    }
-
-    tasks.withType<Detekt> detekt@{
-        jvmTarget = maxKotlinJvmTarget.toString()
-
-        dependsOn(":detekt-rules:assemble")
-
-        reports {
-            html.required.set(false)
-
-            // TODO: Enable this once https://github.com/detekt/detekt/issues/5034 is resolved and use the merged
-            //       Markdown file as a GitHub Action job summary, see
-            //       https://github.blog/2022-05-09-supercharging-github-actions-with-job-summaries/.
-            md.required.set(false)
-
-            sarif.required.set(true)
-            txt.required.set(false)
-            xml.required.set(false)
-        }
-
-        finalizedBy(mergeDetektReports)
-
-        mergeDetektReports.configure {
-            input.from(this@detekt.sarifReportFile)
-        }
-    }
-}
-
-subprojects {
-    version = rootProject.version
-
-    val nonJavaProjects = listOf(
-        "commands",
-        "package-curation-providers",
-        "package-managers",
-        "reporters",
-        "web-app-template"
-    )
-
-    if (name in nonJavaProjects) return@subprojects
-
-    // Apply core plugins.
-    apply(plugin = "jacoco")
-    apply(plugin = "maven-publish")
-
-    // Apply third-party plugins.
-    apply(plugin = "org.jetbrains.kotlin.jvm")
-    apply(plugin = "dev.adamko.dokkatoo")
-
-    testing {
-        suites {
-            withType<JvmTestSuite>().configureEach {
-                useJUnitJupiter()
-
-                dependencies {
-                    implementation(project(":utils:test-utils"))
-
-                    implementation(rootProject.libs.kotestAssertionsCore)
-                    implementation(rootProject.libs.kotestRunnerJunit5)
-                }
-            }
-
-            register<JvmTestSuite>("funTest") {
-                sources {
-                    kotlin {
-                        testType.set(TestSuiteType.FUNCTIONAL_TEST)
-                    }
-                }
-            }
-        }
-    }
-
-    // Associate the "funTest" compilation with the "main" compilation to be able to access "internal" objects from
-    // functional tests.
-    kotlin.target.compilations.apply {
-        getByName("funTest").associateWith(getByName(KotlinCompilation.MAIN_COMPILATION_NAME))
-    }
-
-    configurations.all {
-        resolutionStrategy {
-            // Ensure all OkHttp versions match our version >= 4 to avoid Kotlin vs. Java issues with OkHttp 3.
-            force(rootProject.libs.okhttp)
-
-            // Ensure all JRuby versions match our version to avoid Psych YAML library issues.
-            force(rootProject.libs.jruby)
-
-            // Ensure all Log4j API versions match our version.
-            force(rootProject.libs.log4jApi)
-
-            // Ensure that all transitive versions of Kotlin libraries match our version of Kotlin.
-            force("org.jetbrains.kotlin:kotlin-reflect:${rootProject.libs.versions.kotlinPlugin.get()}")
-        }
-    }
-
-    tasks.withType<JavaCompile>().configureEach {
-        // Align this with Kotlin to avoid errors, see https://youtrack.jetbrains.com/issue/KT-48745.
-        sourceCompatibility = maxKotlinJvmTarget.toString()
-        targetCompatibility = maxKotlinJvmTarget.toString()
-    }
-
-    tasks.withType<KotlinCompile>().configureEach {
-        val customCompilerArgs = listOf(
-            "-opt-in=kotlin.contracts.ExperimentalContracts",
-            "-opt-in=kotlin.io.path.ExperimentalPathApi",
-            "-opt-in=kotlin.time.ExperimentalTime"
-        )
-
-        kotlinOptions {
-            allWarningsAsErrors = true
-            apiVersion = "1.8"
-            freeCompilerArgs = freeCompilerArgs + customCompilerArgs
-            jvmTarget = maxKotlinJvmTarget.toString()
-        }
-    }
-
-    tasks.withType<Test>().configureEach {
-        // Work-around for "--tests" only being able to include tests, see https://github.com/gradle/gradle/issues/6505.
-        properties["tests.exclude"]?.also { excludes ->
-            filter {
-                excludes.toString().split(',').map { excludeTestsMatching(it) }
-                isFailOnNoMatchingTests = false
-            }
-        }
-
-        // Convenience alternative to "--tests" that can take multiple patterns at once as Gradle is not planning to
-        // implement this, see https://github.com/gradle/gradle/issues/5719.
-        properties["tests.include"]?.also { includes ->
-            filter {
-                includes.toString().split(',').map { includeTestsMatching(it) }
-                isFailOnNoMatchingTests = false
-            }
-        }
-
-        if (javaVersion.isCompatibleWith(JavaVersion.VERSION_17)) {
-            // See https://kotest.io/docs/next/extensions/system_extensions.html#system-environment.
-            jvmArgs(
-                "--add-opens", "java.base/java.io=ALL-UNNAMED",
-                "--add-opens", "java.base/java.util=ALL-UNNAMED",
-                "--add-opens", "java.base/sun.nio.ch=ALL-UNNAMED"
-            )
-        }
-
-        val testSystemProperties = mutableListOf("gradle.build.dir" to project.buildDir.path)
-
-        listOf(
-            "java.io.tmpdir",
-            "kotest.assertions.multi-line-diff",
-            "kotest.tags"
-        ).mapNotNullTo(testSystemProperties) { key ->
-            System.getProperty(key)?.let { key to it }
-        }
-
-        systemProperties = testSystemProperties.toMap()
-
-        testLogging {
-            events = setOf(TestLogEvent.STARTED, TestLogEvent.PASSED, TestLogEvent.SKIPPED, TestLogEvent.FAILED)
-            exceptionFormat = TestExceptionFormat.FULL
-            showCauses = false
-            showStackTraces = false
-            showStandardStreams = false
-        }
-    }
-
-    tasks.named<JacocoReport>("jacocoTestReport") {
-        reports {
-            // Enable XML in addition to HTML for CI integration.
-            xml.required.set(true)
-        }
-    }
-
-    tasks.register<JacocoReport>("jacocoFunTestReport") {
-        description = "Generates code coverage report for the funTest task."
-        group = "Reporting"
-
-        executionData(tasks["funTest"])
-        sourceSets(sourceSets.main.get())
-
-        reports {
-            // Enable XML in addition to HTML for CI integration.
-            xml.required.set(true)
-        }
-    }
-
-    tasks.register("jacocoReport") {
-        description = "Generates code coverage reports for all test tasks."
-        group = "Reporting"
-
-        dependsOn(tasks.withType<JacocoReport>())
-    }
-
-    tasks.named("check") {
-        dependsOn(tasks["funTest"])
-    }
-
-    tasks.withType<Jar>().configureEach {
-        duplicatesStrategy = DuplicatesStrategy.EXCLUDE
-        isPreserveFileTimestamps = false
-        isReproducibleFileOrder = true
-
-        manifest {
-            attributes["Implementation-Version"] = project.version
-        }
-    }
-
-    tasks.register<Jar>("sourcesJar") {
-        archiveClassifier.set("sources")
-        from(sourceSets.main.get().allSource)
-    }
-
-    tasks.register<Jar>("docsHtmlJar") {
-        description = "Assembles a JAR containing the HTML documentation."
-        group = "Documentation"
-
-        dependsOn(tasks.dokkatooGeneratePublicationHtml)
-        from(tasks.dokkatooGeneratePublicationHtml.flatMap { it.outputDirectory })
-        archiveClassifier.set("htmldoc")
-    }
-
-    tasks.register<Jar>("docsJavadocJar") {
-        description = "Assembles a JAR containing the Javadoc documentation."
-        group = "Documentation"
-
-        dependsOn(tasks.dokkatooGeneratePublicationJavadoc)
-        from(tasks.dokkatooGeneratePublicationJavadoc.flatMap { it.outputDirectory })
-        archiveClassifier.set("javadoc")
-    }
-
-    configure<PublishingExtension> {
-        publications {
-            create<MavenPublication>(name) {
-                fun getGroupId(parent: Project?): String =
-                    if (parent == null) "" else "${getGroupId(parent.parent)}.${parent.name.replace("-", "")}"
-
-                groupId = "org${getGroupId(parent)}"
-
-                from(components["java"])
-                artifact(tasks["sourcesJar"])
-                artifact(tasks["docsJavadocJar"])
-
-                pom {
-                    licenses {
-                        license {
-                            name.set("Apache-2.0")
-                            url.set("https://www.apache.org/licenses/LICENSE-2.0")
-                        }
-                    }
-
-                    scm {
-                        connection.set("scm:git:https://github.com/oss-review-toolkit/ort.git")
-                        developerConnection.set("scm:git:git@github.com:oss-review-toolkit/ort.git")
-                        tag.set(version.toString())
-                        url.set("https://github.com/oss-review-toolkit/ort")
-                    }
-                }
-            }
-        }
     }
 }
