+++ plugins/package-managers/gradle-model/build.gradle.kts
@@ -17,4 +17,7 @@
  * License-Filename: LICENSE
  */
 
-// Nothing needed here.
+plugins {
+    // Apply precompiled plugins.
+    id("ort-kotlin-conventions")
+}
